NEW FEATURE & BUGFIX : Make the context menu (right-click menu) dynamic. Now if you click a file, a folder, or the empty background in the file list view, you'll get a different more focused menu showing you only the tasks you can do with the selected item. Cool! This is because of a bug where you right-click when a file is not selected then DELETE and RENAME should greyed out on the Context Menu (right-click menu) but they don't. The toolbar now hides DELETE and RENAME when it should as well.
ADDED: Menu item in help for opening the Licence.txt file.

TWEAK: Made the Text Viewer just a little bit wider by default since that seemed to handle the most of the text file I looked at with it. I went through a bunch of Amiga TPUG disk images to test this which was an amazing resource!

TWEAK: Long file lists create scroll bars again. So make the fields a little tighter so they all fit without a scroll!

TWEAK: Added line breaks in between the bullet list items in the Readme.txt file. Also add a line in the readme file that the help file is not updated yet.

TWEAK: Preferences dialog box --> Added more padding and tweaking.